+++ docs/html/search/all_e.js
@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['tm1637module_57',['Tm1637Module',['../classace__segment_1_1Tm1637Module.html',1,'ace_segment::Tm1637Module&lt; T_TMII, T_DIGITS &gt;'],['../classace__segment_1_1Tm1637Module.html#a9cc401ceeed8292b26a93d5dcc08312e',1,'ace_segment::Tm1637Module::Tm1637Module()']]],
-  ['tm1638anodemodule_58',['Tm1638AnodeModule',['../classace__segment_1_1Tm1638AnodeModule.html',1,'ace_segment::Tm1638AnodeModule&lt; T_TMII, T_DIGITS &gt;'],['../classace__segment_1_1Tm1638AnodeModule.html#aa981626f917c6e7bacec2e2643042797',1,'ace_segment::Tm1638AnodeModule::Tm1638AnodeModule()']]],
-  ['tm1638module_59',['Tm1638Module',['../classace__segment_1_1Tm1638Module.html',1,'ace_segment::Tm1638Module&lt; T_TMII, T_DIGITS &gt;'],['../classace__segment_1_1Tm1638Module.html#a0fdb90f996a0f47838a76bc500550fd4',1,'ace_segment::Tm1638Module::Tm1638Module()']]]
+  ['tm1637module_59',['Tm1637Module',['../classace__segment_1_1Tm1637Module.html',1,'ace_segment::Tm1637Module&lt; T_TMII, T_DIGITS &gt;'],['../classace__segment_1_1Tm1637Module.html#a9cc401ceeed8292b26a93d5dcc08312e',1,'ace_segment::Tm1637Module::Tm1637Module()']]],
+  ['tm1638anodemodule_60',['Tm1638AnodeModule',['../classace__segment_1_1Tm1638AnodeModule.html',1,'ace_segment::Tm1638AnodeModule&lt; T_TMII, T_DIGITS &gt;'],['../classace__segment_1_1Tm1638AnodeModule.html#aa981626f917c6e7bacec2e2643042797',1,'ace_segment::Tm1638AnodeModule::Tm1638AnodeModule()']]],
+  ['tm1638module_61',['Tm1638Module',['../classace__segment_1_1Tm1638Module.html',1,'ace_segment::Tm1638Module&lt; T_TMII, T_DIGITS &gt;'],['../classace__segment_1_1Tm1638Module.html#a0fdb90f996a0f47838a76bc500550fd4',1,'ace_segment::Tm1638Module::Tm1638Module()']]]
 ];