+++ docs/html/search/classes_4.js
@@ -1,10 +1,10 @@
 var searchData=
 [
-  ['ledmatrixbase_67',['LedMatrixBase',['../classace__segment_1_1LedMatrixBase.html',1,'ace_segment']]],
-  ['ledmatrixdirect_68',['LedMatrixDirect',['../classace__segment_1_1LedMatrixDirect.html',1,'ace_segment']]],
-  ['ledmatrixdirect_3c_20gpiointerface_20_3e_69',['LedMatrixDirect&lt; GpioInterface &gt;',['../classace__segment_1_1LedMatrixDirect.html',1,'ace_segment']]],
-  ['ledmatrixdualhc595_70',['LedMatrixDualHc595',['../classace__segment_1_1LedMatrixDualHc595.html',1,'ace_segment']]],
-  ['ledmatrixsinglehc595_71',['LedMatrixSingleHc595',['../classace__segment_1_1LedMatrixSingleHc595.html',1,'ace_segment']]],
-  ['ledmatrixsinglehc595_3c_20t_5fspii_2c_20gpiointerface_20_3e_72',['LedMatrixSingleHc595&lt; T_SPII, GpioInterface &gt;',['../classace__segment_1_1LedMatrixSingleHc595.html',1,'ace_segment']]],
-  ['ledmodule_73',['LedModule',['../classace__segment_1_1LedModule.html',1,'ace_segment']]]
+  ['ledmatrixbase_69',['LedMatrixBase',['../classace__segment_1_1LedMatrixBase.html',1,'ace_segment']]],
+  ['ledmatrixdirect_70',['LedMatrixDirect',['../classace__segment_1_1LedMatrixDirect.html',1,'ace_segment']]],
+  ['ledmatrixdirect_3c_20gpiointerface_20_3e_71',['LedMatrixDirect&lt; GpioInterface &gt;',['../classace__segment_1_1LedMatrixDirect.html',1,'ace_segment']]],
+  ['ledmatrixdualhc595_72',['LedMatrixDualHc595',['../classace__segment_1_1LedMatrixDualHc595.html',1,'ace_segment']]],
+  ['ledmatrixsinglehc595_73',['LedMatrixSingleHc595',['../classace__segment_1_1LedMatrixSingleHc595.html',1,'ace_segment']]],
+  ['ledmatrixsinglehc595_3c_20t_5fspii_2c_20gpiointerface_20_3e_74',['LedMatrixSingleHc595&lt; T_SPII, GpioInterface &gt;',['../classace__segment_1_1LedMatrixSingleHc595.html',1,'ace_segment']]],
+  ['ledmodule_75',['LedModule',['../classace__segment_1_1LedModule.html',1,'ace_segment']]]
 ];