+++ docs/html/search/classes_7.js
@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['tm1637module_80',['Tm1637Module',['../classace__segment_1_1Tm1637Module.html',1,'ace_segment']]],
-  ['tm1638anodemodule_81',['Tm1638AnodeModule',['../classace__segment_1_1Tm1638AnodeModule.html',1,'ace_segment']]],
-  ['tm1638module_82',['Tm1638Module',['../classace__segment_1_1Tm1638Module.html',1,'ace_segment']]]
+  ['tm1637module_82',['Tm1637Module',['../classace__segment_1_1Tm1637Module.html',1,'ace_segment']]],
+  ['tm1638anodemodule_83',['Tm1638AnodeModule',['../classace__segment_1_1Tm1638AnodeModule.html',1,'ace_segment']]],
+  ['tm1638module_84',['Tm1638Module',['../classace__segment_1_1Tm1638Module.html',1,'ace_segment']]]
 ];