+++ docs/html/search/functions_5.js
@@ -1,10 +1,10 @@
 var searchData=
 [
-  ['getbrightness_96',['getBrightness',['../classace__segment_1_1LedModule.html#aed70d70812a3f5d1335b0073f6542afb',1,'ace_segment::LedModule']]],
-  ['getfieldsperframe_97',['getFieldsPerFrame',['../classace__segment_1_1ScanningModule.html#ad9388b85594d5a74038045b642938286',1,'ace_segment::ScanningModule']]],
-  ['getfieldspersecond_98',['getFieldsPerSecond',['../classace__segment_1_1ScanningModule.html#a11de5fb5716aa5650406994ca827a7c7',1,'ace_segment::ScanningModule']]],
-  ['getframespersecond_99',['getFramesPerSecond',['../classace__segment_1_1ScanningModule.html#a8a41adc98b3e608923975e432e4e5683',1,'ace_segment::ScanningModule']]],
-  ['getmicrosperfield_100',['getMicrosPerField',['../classace__segment_1_1ScanningModule.html#a8ff2aad4bd53b7da4490c6e349e8b0a2',1,'ace_segment::ScanningModule']]],
-  ['getnumdigits_101',['getNumDigits',['../classace__segment_1_1LedModule.html#a5b458f082ece47f08d0640ad3168cab1',1,'ace_segment::LedModule']]],
-  ['getpatternat_102',['getPatternAt',['../classace__segment_1_1LedModule.html#a50688891c17deead2fb2075a0d88c189',1,'ace_segment::LedModule']]]
+  ['getbrightness_98',['getBrightness',['../classace__segment_1_1LedModule.html#aed70d70812a3f5d1335b0073f6542afb',1,'ace_segment::LedModule']]],
+  ['getfieldsperframe_99',['getFieldsPerFrame',['../classace__segment_1_1ScanningModule.html#ad9388b85594d5a74038045b642938286',1,'ace_segment::ScanningModule']]],
+  ['getfieldspersecond_100',['getFieldsPerSecond',['../classace__segment_1_1ScanningModule.html#a11de5fb5716aa5650406994ca827a7c7',1,'ace_segment::ScanningModule']]],
+  ['getframespersecond_101',['getFramesPerSecond',['../classace__segment_1_1ScanningModule.html#a8a41adc98b3e608923975e432e4e5683',1,'ace_segment::ScanningModule']]],
+  ['getmicrosperfield_102',['getMicrosPerField',['../classace__segment_1_1ScanningModule.html#a8ff2aad4bd53b7da4490c6e349e8b0a2',1,'ace_segment::ScanningModule']]],
+  ['getnumdigits_103',['getNumDigits',['../classace__segment_1_1LedModule.html#a5b458f082ece47f08d0640ad3168cab1',1,'ace_segment::LedModule']]],
+  ['getpatternat_104',['getPatternAt',['../classace__segment_1_1LedModule.html#a50688891c17deead2fb2075a0d88c189',1,'ace_segment::LedModule']]]
 ];