+++ docs/html/search/functions_6.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['hc595module_103',['Hc595Module',['../classace__segment_1_1Hc595Module.html#ae3ce4d5d32c92c01df7a424c6e5e56cc',1,'ace_segment::Hc595Module']]],
-  ['ht16k33module_104',['Ht16k33Module',['../classace__segment_1_1Ht16k33Module.html#af36ff9ba6dffb7c445cdf0b4c6a1761f',1,'ace_segment::Ht16k33Module']]]
+  ['hc595module_105',['Hc595Module',['../classace__segment_1_1Hc595Module.html#ae3ce4d5d32c92c01df7a424c6e5e56cc',1,'ace_segment::Hc595Module']]],
+  ['ht16k33module_106',['Ht16k33Module',['../classace__segment_1_1Ht16k33Module.html#af36ff9ba6dffb7c445cdf0b4c6a1761f',1,'ace_segment::Ht16k33Module']]]
 ];