+++ docs/html/search/functions_b.js
@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['readbuttons_118',['readButtons',['../classace__segment_1_1Tm1637Module.html#a73fb704b87858583de50f61e52ff205a',1,'ace_segment::Tm1637Module::readButtons()'],['../classace__segment_1_1Tm1638AnodeModule.html#a7a7b28d3eb12629b522fd51a747a1ecf',1,'ace_segment::Tm1638AnodeModule::readButtons()'],['../classace__segment_1_1Tm1638Module.html#a8b8404c5c87a908aa2f81ef7af892d4b',1,'ace_segment::Tm1638Module::readButtons()']]],
-  ['renderfieldnow_119',['renderFieldNow',['../classace__segment_1_1ScanningModule.html#ac6a00935339bb99f7a13e514a7252c77',1,'ace_segment::ScanningModule']]],
-  ['renderfieldwhenready_120',['renderFieldWhenReady',['../classace__segment_1_1ScanningModule.html#a31c1fcc0f98a937137752914398de85b',1,'ace_segment::ScanningModule']]]
+  ['readbuttons_120',['readButtons',['../classace__segment_1_1Tm1637Module.html#a73fb704b87858583de50f61e52ff205a',1,'ace_segment::Tm1637Module::readButtons()'],['../classace__segment_1_1Tm1638AnodeModule.html#a7a7b28d3eb12629b522fd51a747a1ecf',1,'ace_segment::Tm1638AnodeModule::readButtons()'],['../classace__segment_1_1Tm1638Module.html#a8b8404c5c87a908aa2f81ef7af892d4b',1,'ace_segment::Tm1638Module::readButtons()']]],
+  ['renderfieldnow_121',['renderFieldNow',['../classace__segment_1_1ScanningModule.html#ac6a00935339bb99f7a13e514a7252c77',1,'ace_segment::ScanningModule']]],
+  ['renderfieldwhenready_122',['renderFieldWhenReady',['../classace__segment_1_1ScanningModule.html#a31c1fcc0f98a937137752914398de85b',1,'ace_segment::ScanningModule']]]
 ];