+++ examples/MemoryBenchmark/generate_readme.py
@@ -38,7 +38,7 @@ by the runtime environment of the processor. For example, it often seems like
 the ESP8266 allocates flash memory in blocks of a certain quantity, so the
 calculated flash size can jump around in unexpected ways.
 
-**Version**: AceSegment v0.12.0
+**Version**: AceSegment v0.13.0
 
 **DO NOT EDIT**: This file was auto-generated using `make README.md`.
 
@@ -252,6 +252,15 @@ before substantional refactoring in 2021.
 
 * Add `Tm1638AnodeModule`. Very similar to `Tm1638Module`.
 
+**v0.13**
+
+* Upgrade tool chain
+    * STM32duino to 2.4.0
+    * ESP32 to 2.0.7
+    * Teensyduino to 1.57
+* Significant increase in flash memory usage for STM32duino (1-4 kB) and ESP32
+  (4-8 kB).
+
 ## Results
 
 The following shows the flash and static memory sizes of the `MemoryBenchmark`
@@ -273,7 +282,7 @@ program for various LED modules.
 ### ATtiny85
 
 * 8MHz ATtiny85
-* Arduino IDE 1.8.19, Arduino CLI 0.19.2
+* Arduino IDE 1.8.19, Arduino CLI 0.31.0
 * SpenceKonde/ATTinyCore 1.5.2
 
 ```
@@ -283,8 +292,8 @@ program for various LED modules.
 ### Arduino Nano
 
 * 16MHz ATmega328P
-* Arduino IDE 1.8.19, Arduino CLI 0.19.2
-* Arduino AVR Boards 1.8.4
+* Arduino IDE 1.8.19, Arduino CLI 0.31.0
+* Arduino AVR Boards 1.8.5
 
 ```
 {nano_results}
@@ -293,7 +302,7 @@ program for various LED modules.
 ### Sparkfun Pro Micro
 
 * 16 MHz ATmega32U4
-* Arduino IDE 1.8.19, Arduino CLI 0.19.2
+* Arduino IDE 1.8.19, Arduino CLI 0.31.0
 * SparkFun AVR Boards 1.1.13
 
 ```
@@ -303,8 +312,8 @@ program for various LED modules.
 ### STM32 Blue Pill
 
 * STM32F103C8, 72 MHz ARM Cortex-M3
-* Arduino IDE 1.8.19, Arduino CLI 0.19.2
-* STM32duino 2.2.0
+* Arduino IDE 1.8.19, Arduino CLI 0.31.0
+* STM32duino 2.4.0
 
 ```
 {stm32_results}
@@ -313,7 +322,7 @@ program for various LED modules.
 ### ESP8266
 
 * NodeMCU 1.0, 80MHz ESP8266
-* Arduino IDE 1.8.19, Arduino CLI 0.19.2
+* Arduino IDE 1.8.19, Arduino CLI 0.31.0
 * ESP8266 Boards 3.0.2
 
 ```
@@ -323,8 +332,8 @@ program for various LED modules.
 ### ESP32
 
 * ESP32-01 Dev Board, 240 MHz Tensilica LX6
-* Arduino IDE 1.8.19, Arduino CLI 0.19.2
-* ESP32 Boards 2.0.2
+* Arduino IDE 1.8.19, Arduino CLI 0.31.0
+* ESP32 Boards 2.0.7
 
 ```
 {esp32_results}
@@ -333,8 +342,8 @@ program for various LED modules.
 ### Teensy 3.2
 
 * 96 MHz ARM Cortex-M4
-* Arduino IDE 1.8.19, Arduino CLI 0.19.2
-* Teensyduino 1.56
+* Arduino IDE 1.8.19, Arduino CLI 0.31.0
+* Teensyduino 1.57
 * Compiler options: "Faster"
 
 ```