+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
 * Unreleased
+* 0.13.0 (2023-03-15)
+    * `LedModule.h`
+        * Add `size()` as alternate form of `getNumDigits()`. Old version
+          retained for backwards compatibility.
+        * Pull `setDecimalPointAt()` from `PatternWriter` down
+          to the `LedModule` layer. It's a low-level operation, seems more
+          appropriate at the `LedModule` layer.
 * 0.12.0 (2022-03-01)
     * Fix invalid pins in `examples/Tm1638Demo` on ESP32 dev board.
     * Add `uint32_t Tm1638Module::readButtons()` method.